+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # Change Log
 
-# Next release (Version TBD)
+# Next release 1.2.0
 
 - File parsing uses the [wellen](https://github.com/ekiwi/wellen/tree/new-api) library compiled to wasm. Benefits include:
   - FST and GHW file support
   - Improves file parsing speed and memory efficiency (over storing everything in JS objects)
 - Removed checkboxes for scope \[module\] items in netlist viewer to reduce confusion
 - Variables loaded into viewer show up before waveform data is loaded as a better visual acknowledgement to user action
+- Scroll Position now limited to end of trace rather than the end of the last chunk
 
 # 1.1.0